setting up & reorganizing into functional categories
