+++ server/src/main/scala/com/ossuminc/riddl/lsp/server/RiddlLSPTextDocumentService/DocContentMgmt/Folding.scala
@@ -1,0 +1,6 @@
+package com.ossuminc.riddl.lsp.server.RiddlLSPTextDocumentService.DocContentMgmt
+
+// foldingRange
+object Folding {
+
+}
