+++ server/src/main/scala/com/ossuminc/riddl/lsp/server/RiddlLSPTextDocumentService/DocContentMgmt/FormattingAndRenaming.scala
@@ -1,0 +1,13 @@
+package com.ossuminc.riddl.lsp.server.RiddlLSPTextDocumentService.DocContentMgmt
+
+/*
+- formatting
+- onTypeFormatting
+- rangeFormatting
+- rename
+- prepareRename
+ */
+
+object FormattingAndRenaming {
+
+}
