+++ server/src/main/scala/com/ossuminc/riddl/lsp/server/RiddlLSPTextDocumentService/DocContentMgmt/LinksAndColours.scala
@@ -1,0 +1,12 @@
+package com.ossuminc.riddl.lsp.server.RiddlLSPTextDocumentService.DocContentMgmt
+
+/*
+- documentLink
+- documentLinkResolve
+- documentColor
+- colorPresentation
+ */
+
+object LinksAndColours {
+  
+}
