+++ server/src/main/scala/com/ossuminc/riddl/lsp/server/RiddlLSPTextDocumentService/DocContentMgmt/SymbolInfo.scala
@@ -1,0 +1,15 @@
+package com.ossuminc.riddl.lsp.server.RiddlLSPTextDocumentService.DocContentMgmt
+
+/*
+- declaration
+- definition
+- typeDefinition
+- implementation
+- references
+- documentHighlight
+- documentSymbol
+ */
+
+object SymbolInfo {
+  
+}
