+++ server/src/main/scala/com/ossuminc/riddl/lsp/server/RiddlLSPTextDocumentService/WorkspaceMgmt.scala
@@ -1,0 +1,3 @@
+package com.ossuminc.riddl.lsp.server.RiddlLSPTextDocumentService
+
+
