+++ server/src/test/scala/com/ossum/riddl/lsp/server/LSPTextDocumentSpec/DocContentMgmtSpec/FoldingSpec.scala
@@ -1,0 +1,6 @@
+package com.ossum.riddl.lsp.server.LSPTextDocumentSpec.DocContentMgmtSpec
+
+// - foldingRange
+class FoldingSpec {
+
+}
