+++ server/src/test/scala/com/ossum/riddl/lsp/server/LSPTextDocumentSpec/DocContentMgmtSpec/FormattingAndRenamingSpec.scala
@@ -1,0 +1,13 @@
+package com.ossum.riddl.lsp.server.LSPTextDocumentSpec.DocContentMgmtSpec
+
+/*
+- formatting
+- onTypeFormatting
+- rangeFormatting
+- rename
+- prepareRename
+ */
+
+class FormattingAndRenamingSpec {
+
+}
