+++ server/src/test/scala/com/ossum/riddl/lsp/server/LSPTextDocumentSpec/DocContentMgmtSpec/LinksAndColoursSpec.scala
@@ -1,0 +1,12 @@
+package com.ossum.riddl.lsp.server.LSPTextDocumentSpec.DocContentMgmtSpec
+
+/*
+- documentLink
+- documentLinkResolve
+- documentColor
+- colorPresentation
+ */
+
+object LinksAndColoursSpec {
+  
+}
