+++ server/src/test/scala/com/ossum/riddl/lsp/server/LSPTextDocumentSpec/DocContentMgmtSpec/SymbolInfoSpec.scala
@@ -1,0 +1,15 @@
+package com.ossum.riddl.lsp.server.LSPTextDocumentSpec.DocContentMgmtSpec
+
+/*
+- declaration
+- definition
+- typeDefinition
+- implementation
+- references
+- documentHighlight
+- documentSymbol
+ */
+
+object SymbolInfoSpec {
+  
+}
