+++ server/src/test/scala/com/ossum/riddl/lsp/server/LSPTextDocumentSpec/DocLifecycleMgmtSpec.scala
@@ -1,9 +1,9 @@
-package com.ossum.riddl.lsp.server
+package com.ossum.riddl.lsp.server.LSPTextDocumentSpec
 
-import com.ossum.riddl.lsp.server.InitializationSpecs.*
+import com.ossum.riddl.lsp.server.initializationSpecs.*
 import org.eclipse.lsp4j
-import org.eclipse.lsp4j.jsonrpc.messages
 import org.eclipse.lsp4j.*
+import org.eclipse.lsp4j.jsonrpc.messages
 import org.scalatest.concurrent.Futures.{whenReady, whenReadyImpl}
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.matchers.must.Matchers
@@ -13,51 +13,16 @@ import java.io.File
 import java.util
 import java.util.concurrent.CompletableFuture
 import scala.concurrent.Future
-import scala.language.postfixOps
 import scala.jdk.CollectionConverters.*
 import scala.jdk.FutureConverters.*
+import scala.language.postfixOps
 
-class RiddlLSPTextDocumentSpec
+class DocLifecycleMgmtSpec
     extends AnyWordSpec
     with Matchers
     with ScalaFutures {
 
-  trait DiagnosticRequestSpec extends DocumentIdentifierSpec {
-    val request = new DocumentDiagnosticParams(textDocumentIdentifier)
-
-    val diagnosticResultF: CompletableFuture[DocumentDiagnosticReport] =
-      service.diagnostic(request)
-  }
-
-  trait CompletionRequestSpec extends DocumentIdentifierSpec {
-
-    var completionResultF: CompletableFuture[
-      messages.Either[util.List[CompletionItem], CompletionList]
-    ] = Future.failed(new Throwable()).asJava.toCompletableFuture
-
-    def requestCompletion(): Unit = {
-      params.setPosition(position)
-
-      completionResultF = service.completion(params)
-    }
-
-    val params = new CompletionParams()
-    params.setTextDocument(textDocumentIdentifier)
-    val position = new Position()
-
-  }
-
   "RiddlLSPTextDocumentService" must {
-    "successfully open everything.riddl, failing to retrieve a completion" in new OpenNoErrorFileSpec
-      with CompletionRequestSpec
-      with DiagnosticRequestSpec {
-      position.setLine(1)
-      position.setCharacter(1)
-      requestCompletion()
-
-      completionResultF.asScala.failed.futureValue mustBe a[Throwable]
-      completionResultF.asScala.failed.futureValue.getMessage mustEqual "Document has no errors"
-    }
 
     "successfully open everythingOneError.riddl, retrieving a completion" in new OpenOneErrorFileSpec
       with CompletionRequestSpec {
@@ -71,17 +36,6 @@ class RiddlLSPTextDocumentSpec
         completion.getRight.getItems.asScala.length mustEqual 1
         completion.getRight.getItems.asScala.head.getDetail mustEqual "Expected one of (end-of-input | whitespace after keyword)"
       }
-    }
-
-    "successfully open empty.riddl, failing to retrieve a completion" in new OpenEmptyFileSpec
-      with CompletionRequestSpec {
-
-      position.setLine(1)
-      position.setCharacter(1)
-      requestCompletion()
-
-      completionResultF.asScala.failed.futureValue mustBe a[Throwable]
-      completionResultF.asScala.failed.futureValue.getMessage mustEqual "Document is empty"
     }
 
     "successfully close everythingOneError.riddl, failing to retrieve a completion" in new OpenOneErrorFileSpec
@@ -120,25 +74,6 @@ class RiddlLSPTextDocumentSpec
         completionList.getRight.getItems.asScala.length mustEqual 1
         completionList.getRight.getItems.asScala.head.getDetail mustEqual
           """Expected one of (end-of-input | whitespace after keyword)"""
-      }
-    }
-
-    "fail requesting for diagnostic from file with no errors" in new OpenNoErrorFileSpec
-      with DiagnosticRequestSpec {
-
-      diagnosticResultF.asScala.failed.futureValue mustBe a[Throwable]
-      diagnosticResultF.asScala.failed.futureValue.getMessage mustEqual "Document has no errors"
-    }
-
-    "succeed requesting for diagnostic from file with one error" in new OpenOneErrorFileSpec
-      with DiagnosticRequestSpec {
-
-      whenReady(diagnosticResultF.asScala) { report =>
-        report.getRelatedFullDocumentDiagnosticReport.getKind mustEqual DocumentDiagnosticReportKind.Full
-        report.getRelatedFullDocumentDiagnosticReport.getItems.asScala.length mustEqual 1
-        report.getRelatedFullDocumentDiagnosticReport.getItems.asScala.head.getMessage mustEqual "Expected one of (end-of-input | whitespace after keyword)"
-        report.getRelatedFullDocumentDiagnosticReport.getItems.asScala.head.getRange.getStart.getLine mustEqual errorLine
-        report.getRelatedFullDocumentDiagnosticReport.getItems.asScala.head.getRange.getStart.getCharacter mustEqual errorCharOnLine
       }
     }
