+++ server/src/test/scala/com/ossum/riddl/lsp/server/LSPTextDocumentSpec/WorkspaceMgmtSpec.scala
@@ -1,0 +1,3 @@
+package com.ossum.riddl.lsp.server.LSPTextDocumentSpec
+
+object WorkspaceMgmtSpec {}
