+++ server/src/test/scala/com/ossum/riddl/lsp/server/requestSpecs.scala
@@ -1,0 +1,10 @@
+package com.ossum.riddl.lsp.server
+
+object requestSpecs {
+  trait DiagnosticRequestSpec extends DocumentIdentifierSpec {
+    val request = new DocumentDiagnosticParams(textDocumentIdentifier)
+
+    val diagnosticResultF: CompletableFuture[DocumentDiagnosticReport] =
+      service.diagnostic(request)
+  }
+}
